feat(server): serve .Rtypes stub buffers with live parse diagnostics
Editing a stub file had zero tooling. Open .Rtypes documents are now
tracked server-side — a rope per buffer, parsed with the stub grammar
against a throwaway interner, errors published per line on open and on
every edit. The engine never sees these buffers: the loaded corpus
stays a set-once input, so the stub-isolation invariant holds.

+++ crates/engine/src/engine.rs
@@ -85,7 +85,7 @@ type MemoTable<K> = RefCell<HashMap<K, Slot<K>>>;
 
 /// The engine's logical clock. Bumped once per [`Engine::set_input`]; memos are validated against it.
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Debug)]
-pub struct Revision(u32);
+pub struct Revision(u64);
 
 impl Revision {
     /// The clock value before any input is set. No real memo is ever verified at `START`, so it is also
